+++ src/main/java/com/oms/orderservice/OrderServiceApplication.java
@@ -1,4 +1,4 @@
-package com.ordermgr.orderservice;
+package com.oms.orderservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
