+++ src/main/java/com/oms/orderservice/util/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.ordermgr.orderservice.util;
+package com.oms.orderservice.util;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
