+++ src/main/java/com/oms/orderservice/util/OrderStatusNameEnum.java
@@ -1,4 +1,4 @@
-package com.ordermgr.orderservice.util;
+package com.oms.orderservice.util;
 
 public enum OrderStatusNameEnum {
