+++ src/main/resources/application.yml
@@ -12,4 +12,4 @@ spring:
 server:
   port: 8082
 api:
-  prefix: /api/v1
+  prefix: /oms/v1
