+++ src/test/java/com/oms/orderservice/OrderServiceImplApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ordermgr.orderservice;
+package com.oms.orderservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
